refactor json rpc methods
