+++ dcrlibwallet.go
@@ -1,15 +1,11 @@
 package dcrlibwallet
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
 	"path/filepath"
-	"strings"
 
 	"github.com/asdine/storm"
-	"github.com/decred/dcrd/dcrjson"
 	"github.com/decred/dcrwallet/errors"
 	"github.com/decred/dcrwallet/netparams"
 	"github.com/decred/dcrwallet/wallet"
@@ -143,76 +139,4 @@ func (lw *LibWallet) Shutdown() {
 			log.Info("tx db closed successfully")
 		}
 	}
-}
-
-func (lw *LibWallet) CallJSONRPC(method string, args string, address string, username string, password string, caCert string) (string, error) {
-	arguments := strings.Split(args, ",")
-	params := make([]interface{}, 0)
-	for _, arg := range arguments {
-		if strings.TrimSpace(arg) == "" {
-			continue
-		}
-		params = append(params, strings.TrimSpace(arg))
-	}
-	// Attempt to create the appropriate command using the arguments
-	// provided by the user.
-	cmd, err := dcrjson.NewCmd(method, params...)
-	if err != nil {
-		// Show the error along with its error code when it's a
-		// dcrjson.Error as it reallistcally will always be since the
-		// NewCmd function is only supposed to return errors of that
-		// type.
-		if jerr, ok := err.(dcrjson.Error); ok {
-			log.Errorf("%s command: %v (code: %s)\n",
-				method, err, jerr.Code)
-			return "", err
-		}
-		// The error is not a dcrjson.Error and this really should not
-		// happen.  Nevertheless, fallback to just showing the error
-		// if it should happen due to a bug in the package.
-		log.Errorf("%s command: %v\n", method, err)
-		return "", err
-	}
-
-	// Marshal the command into a JSON-RPC byte slice in preparation for
-	// sending it to the RPC server.
-	marshalledJSON, err := dcrjson.MarshalCmd("1.0", 1, cmd)
-	if err != nil {
-		log.Error(err)
-		return "", err
-	}
-
-	// Send the JSON-RPC request to the server using the user-specified
-	// connection configuration.
-	result, err := utils.SendPostRequest(marshalledJSON, address, username, password, caCert)
-	if err != nil {
-		log.Error(err)
-		return "", err
-	}
-
-	// Choose how to display the result based on its type.
-	strResult := string(result)
-	if strings.HasPrefix(strResult, "{") || strings.HasPrefix(strResult, "[") {
-		var dst bytes.Buffer
-		if err := json.Indent(&dst, result, "", "  "); err != nil {
-			log.Errorf("Failed to format result: %v", err)
-			return "", err
-		}
-		fmt.Println(dst.String())
-		return dst.String(), nil
-
-	} else if strings.HasPrefix(strResult, `"`) {
-		var str string
-		if err := json.Unmarshal(result, &str); err != nil {
-			log.Errorf("Failed to unmarshal result: %v", err)
-			return "", err
-		}
-		fmt.Println(str)
-		return str, nil
-
-	} else if strResult != "null" {
-		fmt.Println(strResult)
-		return strResult, nil
-	}
-	return "", nil
 }
